+++ packages/amplify-e2e-core/src/utils/sdk-calls.ts
@@ -1,22 +1,4 @@
-import * as path from 'path';
-import {
-  config,
-  DynamoDB,
-  S3,
-  CognitoIdentityServiceProvider,
-  Lambda,
-  LexModelBuildingService,
-  Rekognition,
-  AppSync,
-  CloudWatchLogs,
-  CloudWatchEvents,
-  Kinesis,
-  CloudFormation,
-  AmplifyBackend,
-  IAM,
-  SSM,
-  Location,
-} from 'aws-sdk';
+import { config, DynamoDB, S3, CognitoIdentityServiceProvider, Lambda, AppSync, CloudFormation, AmplifyBackend, IAM } from 'aws-sdk';
 import _ from 'lodash';
 
 export const getDDBTable = async (tableName: string, region: string) => {
@@ -45,30 +27,6 @@ export const bucketNotExists = async (bucket: string) => {
       return false;
     }
     throw error;
-  }
-};
-
-export const getBucketEncryption = async (bucket: string) => {
-  const s3 = new S3();
-  const params = {
-    Bucket: bucket,
-  };
-  try {
-    const result = await s3.getBucketEncryption(params).promise();
-    return result.ServerSideEncryptionConfiguration;
-  } catch (err) {
-    throw new Error(`Error fetching SSE info for bucket ${bucket}. Underlying error was [${err.message}]`);
-  }
-};
-
-export const getBucketKeys = async (params: S3.ListObjectsRequest) => {
-  const s3 = new S3();
-
-  try {
-    const result = await s3.listObjects(params).promise();
-    return result.Contents.map((contentObj) => contentObj.Key);
-  } catch (err) {
-    throw new Error(`Error fetching keys for bucket ${params.Bucket}. Underlying error was [${err.message}]`);
   }
 };
 
@@ -128,14 +86,6 @@ export const getUserPool = async (userpoolId, region) => {
   return res;
 };
 
-export const getMFAConfiguration = async (
-  userPoolId: string,
-  region: string,
-): Promise<CognitoIdentityServiceProvider.GetUserPoolMfaConfigResponse> => {
-  config.update({ region });
-  return await new CognitoIdentityServiceProvider().getUserPoolMfaConfig({ UserPoolId: userPoolId }).promise();
-};
-
 export const getLambdaFunction = async (functionName: string, region: string) => {
   const lambda = new Lambda({ region });
   try {
@@ -164,49 +114,9 @@ export const getUserPoolClients = async (userPoolId: string, clientIds: string[]
   return res;
 };
 
-export const getBot = async (botName: string, region: string) => {
-  const service = new LexModelBuildingService({ region });
-  return await service.getBot({ name: botName, versionOrAlias: '$LATEST' }).promise();
-};
-
-export const getFunction = async (functionName: string, region: string) => {
-  const service = new Lambda({ region });
-  return await service.getFunction({ FunctionName: functionName }).promise();
-};
-
-export const getLayerVersion = async (functionArn: string, region: string) => {
-  const service = new Lambda({ region });
-  return await service.getLayerVersionByArn({ Arn: functionArn }).promise();
-};
-
-export const listVersions = async (layerName: string, region: string) => {
-  const service = new Lambda({ region });
-  return await service.listLayerVersions({ LayerName: layerName }).promise();
-};
-
-export const invokeFunction = async (functionName: string, payload: string, region: string) => {
-  const service = new Lambda({ region });
-  return await service.invoke({ FunctionName: functionName, Payload: payload }).promise();
-};
-
-export const getCollection = async (collectionId: string, region: string) => {
-  const service = new Rekognition({ region });
-  return await service.describeCollection({ CollectionId: collectionId }).promise();
-};
-
 export const getTable = async (tableName: string, region: string) => {
   const service = new DynamoDB({ region });
   return await service.describeTable({ TableName: tableName }).promise();
-};
-
-export const getEventSourceMappings = async (functionName: string, region: string) => {
-  const service = new Lambda({ region });
-  return (await service.listEventSourceMappings({ FunctionName: functionName }).promise()).EventSourceMappings;
-};
-
-export const deleteTable = async (tableName: string, region: string) => {
-  const service = new DynamoDB({ region });
-  return await service.deleteTable({ TableName: tableName }).promise();
 };
 
 export const putItemInTable = async (tableName: string, region: string, item: unknown) => {
@@ -227,25 +137,6 @@ export const getAppSyncApi = async (appSyncApiId: string, region: string) => {
 export const listAppSyncFunctions = async (appSyncApiId: string, region: string) => {
   const service = new AppSync({ region });
   return await service.listFunctions({ apiId: appSyncApiId }).promise();
-};
-
-export const getCloudWatchLogs = async (region: string, logGroupName: string, logStreamName: string | undefined = undefined) => {
-  const cloudwatchlogs = new CloudWatchLogs({ region, retryDelayOptions: { base: 500 } });
-
-  let targetStreamName = logStreamName;
-  if (targetStreamName === undefined) {
-    const describeStreamsResp = await cloudwatchlogs
-      .describeLogStreams({ logGroupName, descending: true, orderBy: 'LastEventTime' })
-      .promise();
-    if (describeStreamsResp.logStreams === undefined || describeStreamsResp.logStreams.length == 0) {
-      return [];
-    }
-
-    targetStreamName = describeStreamsResp.logStreams[0].logStreamName;
-  }
-
-  const logsResp = await cloudwatchlogs.getLogEvents({ logGroupName, logStreamName: targetStreamName }).promise();
-  return logsResp.events || [];
 };
 
 export const describeCloudFormationStack = async (stackName: string, region: string, profileConfig?: any) => {
@@ -287,37 +178,6 @@ export const getTableResourceId = async (region: string, table: string, StackId:
   return null;
 };
 
-export const putKinesisRecords = async (data: string, partitionKey: string, streamName: string, region: string) => {
-  const kinesis = new Kinesis({ region });
-
-  return await kinesis
-    .putRecords({
-      Records: [
-        {
-          Data: data,
-          PartitionKey: partitionKey,
-        },
-      ],
-      StreamName: streamName,
-    })
-    .promise();
-};
-
-export const getCloudWatchEventRule = async (targetName: string, region: string) => {
-  config.update({ region });
-  const service = new CloudWatchEvents();
-  let params = {
-    TargetArn: targetName /* required */,
-  };
-  let ruleName;
-  try {
-    ruleName = await service.listRuleNamesByTarget(params).promise();
-  } catch (e) {
-    console.log(e);
-  }
-  return ruleName;
-};
-
 export const setupAmplifyAdminUI = async (appId: string, region: string) => {
   const amplifyBackend = new AmplifyBackend({ region });
 
@@ -344,69 +204,4 @@ export const listRolePolicies = async (roleName: string, region: string) => {
 export const listAttachedRolePolicies = async (roleName: string, region: string) => {
   const service = new IAM({ region });
   return (await service.listAttachedRolePolicies({ RoleName: roleName }).promise()).AttachedPolicies;
-};
-
-export const getPermissionsBoundary = async (roleName: string, region) => {
-  const iamClient = new IAM({ region });
-  return (await iamClient.getRole({ RoleName: roleName }).promise())?.Role?.PermissionsBoundary?.PermissionsBoundaryArn;
-};
-
-export const getSSMParameters = async (region: string, appId: string, envName: string, funcName: string, parameterNames: string[]) => {
-  const ssmClient = new SSM({ region });
-  if (!parameterNames || parameterNames.length === 0) {
-    throw new Error('no parameterNames specified');
-  }
-  return await ssmClient
-    .getParameters({
-      Names: parameterNames.map((name) => path.posix.join('/amplify', appId, envName, `AMPLIFY_${funcName}_${name}`)),
-      WithDecryption: true,
-    })
-    .promise();
-};
-// Amazon location service calls
-export const getMap = async (mapName: string, region: string) => {
-  const service = new Location({ region });
-  return await service
-    .describeMap({
-      MapName: mapName,
-    })
-    .promise();
-};
-
-export const getPlaceIndex = async (placeIndexName: string, region: string) => {
-  const service = new Location({ region });
-  return await service
-    .describePlaceIndex({
-      IndexName: placeIndexName,
-    })
-    .promise();
-};
-
-export const getGeofenceCollection = async (geofenceCollectionName: string, region: string) => {
-  const service = new Location({ region });
-  return await service
-    .describeGeofenceCollection({
-      CollectionName: geofenceCollectionName,
-    })
-    .promise();
-};
-
-export const getGeofence = async (geofenceCollectionName: string, geofenceId: string, region: string) => {
-  const service = new Location({ region });
-  return (
-    await service.getGeofence({
-      CollectionName: geofenceCollectionName,
-      GeofenceId: geofenceId,
-    })
-  ).promise();
-};
-
-export const listGeofences = async (geofenceCollectionName: string, region: string, nextToken: string = null) => {
-  const service = new Location({ region });
-  return (
-    await service.listGeofences({
-      CollectionName: geofenceCollectionName,
-      NextToken: nextToken,
-    })
-  ).promise();
 };