+++ packages/amplify-graphql-api-construct/src/internal/construct-exports.ts
@@ -14,6 +14,8 @@ import { CfnResource, NestedStack } from 'aws-cdk-lib';
 import { getResourceName } from '@aws-amplify/graphql-transformer-core';
 import { CfnFunction, Function as LambdaFunction } from 'aws-cdk-lib/aws-lambda';
 import { AmplifyGraphqlApiResources, FunctionSlot } from '../types';
+import { AmplifyDynamoDbTableWrapper } from '../amplify-dynamodb-table-wrapper';
+import { walkAndProcessNodes } from './construct-tree';
 
 /**
  * Everything below here is intended to help us gather the
@@ -33,7 +35,7 @@ export const getGeneratedResources = (scope: Construct): AmplifyGraphqlApiResour
   const cfnDataSources: Record<string, CfnDataSource> = {};
   const tables: Record<string, Table> = {};
   const cfnTables: Record<string, CfnTable> = {};
-  const cfnAmplifyTables: Record<string, CfnResource> = {};
+  const amplifyDynamoDbTables: Record<string, AmplifyDynamoDbTableWrapper> = {};
   const roles: Record<string, Role> = {};
   const cfnRoles: Record<string, CfnRole> = {};
   const functions: Record<string, LambdaFunction> = {};
@@ -78,8 +80,8 @@ export const getGeneratedResources = (scope: Construct): AmplifyGraphqlApiResour
       cfnTables[resourceName] = currentScope;
       return;
     }
-    if (currentScope instanceof CfnResource && currentScope.cfnResourceType === 'Custom::AmplifyDynamoDBTable') {
-      cfnAmplifyTables[resourceName] = currentScope;
+    if (AmplifyDynamoDbTableWrapper.isAmplifyDynamoDbTableResource(currentScope)) {
+      amplifyDynamoDbTables[resourceName] = new AmplifyDynamoDbTableWrapper(currentScope);
       return;
     }
     if (currentScope instanceof Role) {
@@ -104,12 +106,7 @@ export const getGeneratedResources = (scope: Construct): AmplifyGraphqlApiResour
     }
   };
 
-  const walkAndClassifyConstructTree = (currentScope: Construct): void => {
-    classifyConstruct(currentScope);
-    currentScope.node.children.forEach(walkAndClassifyConstructTree);
-  };
-
-  scope.node.children.forEach(walkAndClassifyConstructTree);
+  scope.node.children.forEach((child) => walkAndProcessNodes(child, classifyConstruct));
 
   if (!cfnGraphqlApi) {
     throw new Error('Expected to find AWS::AppSync::GraphQLApi in the generated resource scope.');
@@ -126,6 +123,7 @@ export const getGeneratedResources = (scope: Construct): AmplifyGraphqlApiResour
   return {
     graphqlApi: GraphqlApi.fromGraphqlApiAttributes(scope, 'L2GraphqlApi', { graphqlApiId: cfnGraphqlApi.attrApiId }),
     tables,
+    amplifyDynamoDbTables,
     roles,
     functions,
     nestedStacks,
@@ -137,7 +135,6 @@ export const getGeneratedResources = (scope: Construct): AmplifyGraphqlApiResour
       cfnFunctionConfigurations,
       cfnDataSources,
       cfnTables,
-      cfnAmplifyTables,
       cfnRoles,
       cfnFunctions,
       additionalCfnResources,