+++ packages/amplify-graphql-api-construct/src/types.ts
@@ -17,6 +17,7 @@ import { IRole, CfnRole } from 'aws-cdk-lib/aws-iam';
 import { IUserPool } from 'aws-cdk-lib/aws-cognito';
 import { IFunction, CfnFunction } from 'aws-cdk-lib/aws-lambda';
 import { IBucket } from 'aws-cdk-lib/aws-s3';
+import { AmplifyDynamoDbTableWrapper } from './amplify-dynamodb-table-wrapper';
 
 /**
  * Configuration for IAM Authorization on the Graphql Api.
@@ -679,11 +680,6 @@ export interface AmplifyGraphqlApiCfnResources {
   readonly cfnTables: Record<string, CfnTable>;
 
   /**
-   * The Generated Amplify DynamoDB Table L1 Resources, keyed by model name.
-   */
-  readonly cfnAmplifyTables: Record<string, CfnResource>;
-
-  /**
    * The Generated IAM Role L1 Resources, keyed by logicalId.
    */
   readonly cfnRoles: Record<string, CfnRole>;
@@ -713,6 +709,11 @@ export interface AmplifyGraphqlApiResources {
    * The Generated DynamoDB Table L2 Resources, keyed by logicalId.
    */
   readonly tables: Record<string, ITable>;
+
+  /**
+   * The Generated Amplify DynamoDb Table wrapped if produced, keyed by name.
+   */
+  readonly amplifyDynamoDbTables: Record<string, AmplifyDynamoDbTableWrapper>;
 
   /**
    * The Generated IAM Role L2 Resources, keyed by logicalId.