+++ packages/amplify-graphql-model-transformer/src/resources/amplify-dynamodb-table/amplify-dynamo-model-resource-generator.ts
@@ -21,6 +21,7 @@ export class AmplifyDynamoModelResourceGenerator extends DynamoModelResourceGene
   private customResourceServiceToken: string = '';
   private amplifyTableArns: string[] = [];
   private ddbManagerPolicy?: aws_iam.Policy;
+  private customResourceProvider?: custom_resources.Provider;
 
   generateResources(ctx: TransformerContextProvider): void {
     if (!this.isEnabled()) {
@@ -76,7 +77,7 @@ export class AmplifyDynamoModelResourceGenerator extends DynamoModelResourceGene
 
     // lambda that will handle DDB CFN events
     const gsiOnEventHandler = new aws_lambda.Function(scope, ResourceConstants.RESOURCES.TableManagerOnEventHandlerLogicalID, {
-      runtime: aws_lambda.Runtime.NODEJS_16_X,
+      runtime: aws_lambda.Runtime.NODEJS_18_X,
       code: lambdaCode,
       handler: 'amplify-table-manager-handler.onEvent',
       timeout: Duration.minutes(14),
@@ -84,7 +85,7 @@ export class AmplifyDynamoModelResourceGenerator extends DynamoModelResourceGene
 
     // lambda that will poll for provisioning to complete
     const gsiIsCompleteHandler = new aws_lambda.Function(scope, ResourceConstants.RESOURCES.TableManagerIsCompleteHandlerLogicalID, {
-      runtime: aws_lambda.Runtime.NODEJS_16_X,
+      runtime: aws_lambda.Runtime.NODEJS_18_X,
       code: lambdaCode,
       handler: 'amplify-table-manager-handler.isComplete',
       timeout: Duration.minutes(14),