+++ scripts/split-e2e-tests.ts
@@ -2,6 +2,47 @@ import { join } from 'path';
 import * as glob from 'glob';
 import * as fs from 'fs-extra';
 import * as yaml from 'js-yaml';
+
+type ForceTests = 'interactions' | 'containers';
+
+type TestTiming = {
+  test: string;
+  medianRuntime: number;
+};
+
+type ComputeType = 'BUILD_GENERAL1_SMALL' | 'BUILD_GENERAL1_MEDIUM' | 'BUILD_GENERAL1_LARGE';
+
+type BatchBuildJob = {
+  identifier: string;
+  buildspec: string;
+  env: {
+    'compute-type': ComputeType;
+    variables?: [string: string];
+  };
+  'depend-on': string[] | string;
+};
+
+type ConfigBase = {
+  batch: {
+    'build-graph': BatchBuildJob[];
+  };
+  env: {
+    'compute-type': ComputeType;
+    shell: 'bash';
+    variables: [string: string];
+  };
+};
+
+type OSType = 'w' | 'l';
+
+type CandidateJob = {
+  region: string;
+  os: OSType;
+  tests: string[];
+  useParentAccount: boolean;
+  runSolo: boolean;
+};
+
 // Ensure to update packages/amplify-e2e-tests/src/cleanup-e2e-resources.ts is also updated this gets updated
 const AWS_REGIONS_TO_RUN_TESTS = [
   'us-east-1',
@@ -13,13 +54,14 @@ const AWS_REGIONS_TO_RUN_TESTS = [
   'ap-southeast-1',
   'ap-southeast-2',
 ];
+
 // Some services (eg. amazon lex, containers) are not available in all regions
 // Tests added to this list will always run in the specified region
 const FORCE_REGION_MAP = {
   interactions: 'us-west-2',
   containers: 'us-east-1',
 };
-type FORCE_TESTS = 'interactions' | 'containers';
+
 // some tests require additional time, the parent account can handle longer tests (up to 90 minutes)
 const USE_PARENT_ACCOUNT = [
   'src/__tests__/transformer-migrations/searchable-migration',
@@ -36,7 +78,7 @@ const TEST_TIMINGS_PATH = join(REPO_ROOT, 'scripts', 'test-timings.data.json');
 const CODEBUILD_CONFIG_BASE_PATH = join(REPO_ROOT, 'codebuild_specs', 'e2e_workflow_base.yml');
 const CODEBUILD_GENERATE_CONFIG_PATH = join(REPO_ROOT, 'codebuild_specs', 'e2e_workflow.yml');
 const CODEBUILD_DEBUG_CONFIG_PATH = join(REPO_ROOT, 'codebuild_specs', 'debug_workflow.yml');
-const RUN_SOLO = [
+const RUN_SOLO: (string | RegExp)[] = [
   'src/__tests__/apigw.test.ts',
   'src/__tests__/api_2.test.ts',
   'src/__tests__/api_11.test.ts',
@@ -76,70 +118,45 @@ const RUN_SOLO = [
   'src/__tests__/transformer-migrations/model-migration.test.ts',
   'src/__tests__/graphql-v2/searchable-node-to-node-encryption/searchable-previous-deployment-no-node-to-node.test.ts',
   'src/__tests__/graphql-v2/searchable-node-to-node-encryption/searchable-previous-deployment-had-node-to-node.test.ts',
-  // GrapQL E2E tests
+  // GraphQL E2E tests
   'src/__tests__/FunctionTransformerTestsV2.e2e.test.ts',
   'src/__tests__/HttpTransformer.e2e.test.ts',
   'src/__tests__/HttpTransformerV2.e2e.test.ts',
+  // Deploy Velocity tests
+  /src\/__tests__\/deploy-velocity\/.*\.test\.ts/,
 ];
+
 const DEBUG_FLAG = '--debug';
 
-export function loadConfigBase() {
-  return yaml.load(fs.readFileSync(CODEBUILD_CONFIG_BASE_PATH, 'utf8'));
-}
-export function saveConfig(config: any, outputPath: string): void {
-  const output = ['# auto generated file. DO NOT EDIT manually', yaml.dump(config, { noRefs: true })];
-  fs.writeFileSync(outputPath, output.join('\n'));
-}
-export function loadTestTimings(): { timingData: { test: string; medianRuntime: number }[] } {
-  return JSON.parse(fs.readFileSync(TEST_TIMINGS_PATH, 'utf-8'));
-}
-function getTestFiles(dir: string, pattern = 'src/**/*.test.ts'): string[] {
-  return glob.sync(pattern, { cwd: dir });
-}
-type COMPUTE_TYPE = 'BUILD_GENERAL1_MEDIUM' | 'BUILD_GENERAL1_LARGE';
-type BatchBuildJob = {
-  identifier: string;
-  env: {
-    'compute-type': COMPUTE_TYPE;
-    variables: [string: string];
-  };
-};
-type ConfigBase = {
-  batch: {
-    'build-graph': BatchBuildJob[];
-    'fast-fail': boolean;
-  };
-  env: {
-    'compute-type': COMPUTE_TYPE;
-    shell: 'bash';
-    variables: [string: string];
-  };
-};
 const MAX_WORKERS = 4;
-type OS_TYPE = 'w' | 'l';
-type CandidateJob = {
-  region: string;
-  os: OS_TYPE;
-  tests: string[];
-  useParentAccount: boolean;
-  runSolo: boolean;
-};
-const createJob = (os: OS_TYPE, jobIdx: number, runSolo: boolean = false): CandidateJob => {
-  const region = AWS_REGIONS_TO_RUN_TESTS[jobIdx % AWS_REGIONS_TO_RUN_TESTS.length];
-  return {
-    region,
-    os,
-    tests: [],
-    useParentAccount: false,
-    runSolo,
-  };
-};
+
+// eslint-disable-next-line import/namespace
+const loadConfigBase = (): ConfigBase => yaml.load(fs.readFileSync(CODEBUILD_CONFIG_BASE_PATH, 'utf8')) as ConfigBase;
+
+// eslint-disable-next-line import/namespace
+const saveConfig = (config: any, outputPath: string): void =>
+  fs.writeFileSync(outputPath, ['# auto generated file. DO NOT EDIT manually', yaml.dump(config, { noRefs: true })].join('\n'));
+
+// eslint-disable-next-line import/namespace
+const loadTestTimings = (): { timingData: TestTiming[] } => JSON.parse(fs.readFileSync(TEST_TIMINGS_PATH, 'utf-8'));
+
+const getTestFiles = (dir: string, pattern = 'src/**/*.test.ts'): string[] => glob.sync(pattern, { cwd: dir });
+
+const createJob = (os: OSType, jobIdx: number, runSolo = false): CandidateJob => ({
+  region: AWS_REGIONS_TO_RUN_TESTS[jobIdx % AWS_REGIONS_TO_RUN_TESTS.length],
+  os,
+  tests: [],
+  useParentAccount: false,
+  runSolo,
+});
+
 const getTestNameFromPath = (testSuitePath: string): string => {
   const startIndex = testSuitePath.lastIndexOf('/') + 1;
   const endIndex = testSuitePath.lastIndexOf('.test');
   return testSuitePath.substring(startIndex, endIndex).split('.e2e').join('').split('.').join('-');
 };
-const splitTests = (baseJobLinux: any, testDirectory: string, isMigration: boolean, pickTests?: (testSuites: string[]) => string[]) => {
+
+const splitTests = (baseJobLinux: any, testDirectory: string, pickTests?: (testSuites: string[]) => string[]): BatchBuildJob[] => {
   const output: any[] = [];
   let testSuites = getTestFiles(testDirectory);
   if (pickTests && typeof pickTests === 'function') {
@@ -155,12 +172,12 @@ const splitTests = (baseJobLinux: any, testDirectory: string, isMigration: boole
     const runtimeB = testFileRunTimes.find((t: any) => t.test === b)?.medianRuntime ?? 30;
     return runtimeA - runtimeB;
   });
-  const generateJobsForOS = (os: OS_TYPE) => {
+  const generateJobsForOS = (os: OSType): CandidateJob[] => {
     const soloJobs = [];
     let jobIdx = 0;
     const osJobs = [createJob(os, jobIdx)];
     jobIdx++;
-    for (let test of testSuites) {
+    for (const test of testSuites) {
       const currentJob = osJobs[osJobs.length - 1];
 
       const FORCE_REGION = Object.keys(FORCE_REGION_MAP).find((key) => {
@@ -170,12 +187,12 @@ const splitTests = (baseJobLinux: any, testDirectory: string, isMigration: boole
 
       const USE_PARENT = USE_PARENT_ACCOUNT.some((usesParent) => test.startsWith(usesParent));
 
-      if (isMigration || RUN_SOLO.find((solo) => test === solo)) {
+      if (RUN_SOLO.find((solo) => test === solo || test.match(solo))) {
         const newSoloJob = createJob(os, jobIdx, true);
         jobIdx++;
         newSoloJob.tests.push(test);
         if (FORCE_REGION) {
-          newSoloJob.region = FORCE_REGION_MAP[FORCE_REGION as FORCE_TESTS];
+          newSoloJob.region = FORCE_REGION_MAP[FORCE_REGION as ForceTests];
         }
         if (USE_PARENT) {
           newSoloJob.useParentAccount = true;
@@ -187,7 +204,7 @@ const splitTests = (baseJobLinux: any, testDirectory: string, isMigration: boole
       // add the test
       currentJob.tests.push(test);
       if (FORCE_REGION) {
-        currentJob.region = FORCE_REGION_MAP[FORCE_REGION as FORCE_TESTS];
+        currentJob.region = FORCE_REGION_MAP[FORCE_REGION as ForceTests];
       }
       if (USE_PARENT) {
         currentJob.useParentAccount = true;
@@ -202,21 +219,14 @@ const splitTests = (baseJobLinux: any, testDirectory: string, isMigration: boole
     return [...osJobs, ...soloJobs];
   };
   const linuxJobs = generateJobsForOS('l');
-  const getIdentifier = (os: string, names: string) => {
-    let jobName = `${names.replace(/-/g, '_')}`.substring(0, 127);
-    if (isMigration) {
-      const startIndex = baseJobLinux.identifier.lastIndexOf('_');
-      jobName = jobName + baseJobLinux.identifier.substring(startIndex);
-    }
-    return jobName;
-  };
+  const getIdentifier = (names: string): string => `${names.replace(/-/g, '_')}`.substring(0, 127);
   const result: any[] = [];
   linuxJobs.forEach((j) => {
     if (j.tests.length !== 0) {
       const names = j.tests.map((tn) => getTestNameFromPath(tn)).join('_');
       const tmp = {
         ...JSON.parse(JSON.stringify(baseJobLinux)), // deep clone base job
-        identifier: getIdentifier(j.os, names),
+        identifier: getIdentifier(names),
       };
       tmp.env.variables = {};
       tmp.env.variables.TEST_SUITE = j.tests.join('|');
@@ -232,69 +242,71 @@ const splitTests = (baseJobLinux: any, testDirectory: string, isMigration: boole
   });
   return result;
 };
-function main(): void {
+
+const main = (): void => {
   const filteredTests = process.argv.slice(2);
-  const configBase: any = loadConfigBase();
+  const configBase: ConfigBase = loadConfigBase();
   const baseBuildGraph = configBase.batch['build-graph'];
-  const splitE2ETests = splitTests(
-    {
-      identifier: 'run_e2e_tests',
-      buildspec: 'codebuild_specs/run_e2e_tests.yml',
-      env: {
-        'compute-type': 'BUILD_GENERAL1_MEDIUM',
+
+  let builds = [
+    ...splitTests(
+      {
+        identifier: 'run_e2e_tests',
+        buildspec: 'codebuild_specs/run_e2e_tests.yml',
+        env: {
+          'compute-type': 'BUILD_GENERAL1_MEDIUM',
+        },
+        'depend-on': ['publish_to_local_registry'],
       },
-      'depend-on': ['publish_to_local_registry'],
-    },
-    join(REPO_ROOT, 'packages', 'amplify-e2e-tests'),
-    false,
-  );
-  const splitConstructTests = splitTests(
-    {
-      identifier: 'run_cdk_tests',
-      buildspec: 'codebuild_specs/run_cdk_tests.yml',
-      env: {
-        'compute-type': 'BUILD_GENERAL1_MEDIUM',
+      join(REPO_ROOT, 'packages', 'amplify-e2e-tests'),
+    ),
+    ...splitTests(
+      {
+        identifier: 'run_cdk_tests',
+        buildspec: 'codebuild_specs/run_cdk_tests.yml',
+        env: {
+          'compute-type': 'BUILD_GENERAL1_MEDIUM',
+        },
+        'depend-on': ['publish_to_local_registry'],
       },
-      'depend-on': ['publish_to_local_registry'],
-    },
-    join(REPO_ROOT, 'packages', 'amplify-graphql-api-construct-tests'),
-    false,
-  );
-  const splitGqlTests = splitTests(
-    {
-      identifier: 'gql_e2e_tests',
-      buildspec: 'codebuild_specs/graphql_e2e_tests.yml',
-      env: {
-        'compute-type': 'BUILD_GENERAL1_MEDIUM',
+      join(REPO_ROOT, 'packages', 'amplify-graphql-api-construct-tests'),
+    ),
+    ...splitTests(
+      {
+        identifier: 'gql_e2e_tests',
+        buildspec: 'codebuild_specs/graphql_e2e_tests.yml',
+        env: {
+          'compute-type': 'BUILD_GENERAL1_MEDIUM',
+        },
+        'depend-on': ['publish_to_local_registry'],
       },
-      'depend-on': ['publish_to_local_registry'],
-    },
-    join(REPO_ROOT, 'packages', 'graphql-transformers-e2e-tests'),
-    false,
-  );
-  let outputPath = CODEBUILD_GENERATE_CONFIG_PATH;
-  let allBuilds = [...splitE2ETests, ...splitConstructTests, ...splitGqlTests];
+      join(REPO_ROOT, 'packages', 'graphql-transformers-e2e-tests'),
+    ),
+  ];
+
   if (filteredTests.length > 0) {
-    allBuilds = allBuilds.filter((build) => filteredTests.includes(build.identifier));
+    builds = builds.filter((build) => filteredTests.includes(build.identifier));
     if (filteredTests.includes(DEBUG_FLAG)) {
-      allBuilds = allBuilds.map((build) => {
-        return { ...build, 'debug-session': true };
-      });
-      outputPath = CODEBUILD_DEBUG_CONFIG_PATH;
+      builds = builds.map((build) => ({ ...build, 'debug-session': true }));
     }
   }
-  const cleanupResources = {
+
+  const cleanupResources: BatchBuildJob = {
     identifier: 'cleanup_e2e_resources',
     buildspec: 'codebuild_specs/cleanup_e2e_resources.yml',
     env: {
       'compute-type': 'BUILD_GENERAL1_SMALL',
     },
-    'depend-on': allBuilds.length > 0 ? [allBuilds[0].identifier] : 'publish_to_local_registry',
+    'depend-on': builds.length > 0 ? [builds[0].identifier] : 'publish_to_local_registry',
   };
-  console.log(`Total number of splitted jobs: ${allBuilds.length}`);
-  let currentBatch = [...baseBuildGraph, ...allBuilds, cleanupResources];
+
+  console.log(`Total number of splitted jobs: ${builds.length}`);
+  const currentBatch = [...baseBuildGraph, ...builds, cleanupResources];
   configBase.batch['build-graph'] = currentBatch;
+
+  const outputPath = filteredTests.includes(DEBUG_FLAG) ? CODEBUILD_DEBUG_CONFIG_PATH : CODEBUILD_GENERATE_CONFIG_PATH;
   saveConfig(configBase, outputPath);
   console.log(`Successfully generated the buildspec at ${outputPath}`);
-}
+};
+
 main();