+++ packages/amplify-graphql-schema-generator/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.8](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-generator@0.4.7...@aws-amplify/graphql-schema-generator@0.4.8) (2023-10-21)
+
+**Note:** Version bump only for package @aws-amplify/graphql-schema-generator
+
 ## [0.4.7](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-generator@0.4.6...@aws-amplify/graphql-schema-generator@0.4.7) (2023-10-12)
 
 **Note:** Version bump only for package @aws-amplify/graphql-schema-generator